backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
numpy>=1.26.0
scipy>=1.11.0
triangle>=20230923
python-multipart>=0.0.6
shapely>=2.0.0
numba>=0.58.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
httpx>=0.25.0
slowapi>=0.1.9